+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
  - Implemented arithmetic operations for `Transform2D`
  - Completed and improved documentation
 - **breaking**: Removed methods `multipy` from `Transform2D`, since arithmetic operations are defined for `Transform2D` now.
+- **breaking**: Renamed `Transform2D` methods `inversed` to `inverse` and `inverse` to `invert`.
 
 ## [0.9.2] - 2024-06-27